// app/src/generated/models/occupancyPeriodResponseEndReason.ts
/**
 * Generated by orval v8.5.1 🍺
 * Do not edit manually.
 * Buurman API
 * Property Management API for Small Landlords
 * OpenAPI spec version: 0.1.0
 */

/**
 * Reason for ending the self-occupancy period
 * @nullable
 */
export type OccupancyPeriodResponseEndReason =
  | (typeof OccupancyPeriodResponseEndReason)[keyof typeof OccupancyPeriodResponseEndReason]
  | null;

export const OccupancyPeriodResponseEndReason = {
  CONVERTING_TO_RENTAL: 'CONVERTING_TO_RENTAL',
  SELLING: 'SELLING',
  RENOVATION: 'RENOVATION',
  OTHER: 'OTHER',
} as const;