// app/src/generated/models/occupancyPeriodResponseType.ts
/**
 * Generated by orval v8.5.1 🍺
 * Do not edit manually.
 * Buurman API
 * Property Management API for Small Landlords
 * OpenAPI spec version: 0.1.0
 */

/**
 * Nature of the self-occupancy period
 */
export type OccupancyPeriodResponseType =
  (typeof OccupancyPeriodResponseType)[keyof typeof OccupancyPeriodResponseType];

export const OccupancyPeriodResponseType = {
  PERSONAL: 'PERSONAL',
  FAMILY: 'FAMILY',
  BUSINESS: 'BUSINESS',
} as const;